added the source file for task 11

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,11 +8,11 @@
 
 int main(void)
 {
-	double largest;
+	long largest;
 	
-	double i;
+	long i;
 
-	double n = 612852475143;
+	long n = 612852475143;
 
 	for (i = 2; i < n; i++)
 	{
